Compare species annotations across CAMDA, NCBI, and antibiogram datasets

This function compares species annotations from the CAMDA dataset against metadata from NCBI and antibiograms. It returns a dataset that includes consistency status between CAMDA and the external sources, as well as a comparison between the external sources themselves.

Usage

compare_sp(metadata_db = NULL, sra_metadata_db = NULL, antibiograms_db = NULL)

Arguments

metadata_db

A data frame containing CAMDA species annotations. Must include the columns: accession and scientific_name_CAMDA. Example datasets included in the package: training_metadata_db, testing_metadata_cleaned.

sra_metadata_db

A data frame containing metadata from NCBI. Must include the columns: accession and scientific_name_NCBI. A sample dataset is preloaded in the package as sra_metadata_db.

antibiograms_db

A data frame containing species annotations from antibiograms. Must include the columns: accession and scientific_name_Antibiogram. A sample dataset is preloaded in the package as antibiograms_db.

Value

A data frame with the following columns:

accession: Unique sample identifier.

scientific_name_CAMDA: Species annotation from the CAMDA dataset.

scientific_name_NCBI: Species annotation from NCBI metadata.

scientific_name_Antibiogram: Species annotation from antibiogram metadata.

status: Consistency between CAMDA and external sources, with values:

Matched_NCBI_Antibiogram

CAMDA matches both NCBI and antibiogram.

Matched_NCBI

CAMDA matches only NCBI.

Matched_Antibiogram

CAMDA matches only the antibiogram.

Mismatch_with_CAMDA

At least one source has data, but CAMDA does not match either.

Missing_All

No data available from NCBI or antibiogram.

status_reference: Consistency between NCBI and antibiogram sources, with values:

Good_sources

Both sources are present and agree.

Verify_sources

At least one source is present and sources disagree or are partially missing.

No_sources

Both NCBI and antibiogram are missing.

Examples

result <- compare_sp(metadata_db = training_metadata_db, sra_metadata_db, antibiograms_db)